protect ui pages with next-auth session
