+++ src/app/api/auth/[...nextauth]/route.ts
@@ -4,7 +4,7 @@ import User from "@/models/User";
 import NextAuth, { AuthOptions } from "next-auth";
 import CredentialsProvider from "next-auth/providers/credentials";
 
-const authOptions: AuthOptions = {
+export const authOptions: AuthOptions = {
   providers: [
     CredentialsProvider({
       name: "Credentials",
